requirements.txt
Django==5.0.7
psycopg[binary]==3.2.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
Pillow==11.0.0

django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-widget-tweaks==1.5.0
django-filter==24.3
django-import-export==4.1.1

openpyxl==3.1.2
reportlab==4.0.9
python-dateutil==2.9.0.post0

celery==5.3.6
redis==5.0.1

django-celery-beat==2.6.0
django-celery-results==2.5.1

gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==1.2.0
boto3==1.26.118
django-storages==1.13.2